+++ app/src/types/cartItem.ts
@@ -1,0 +1,8 @@
+export type CardItem = {
+  id: string;
+  image: typeof import('*.png');
+  title: string;
+  size: string;
+  quantity: number;
+  unitPrice: number;
+};
